Move override_suse.py to _internal
